requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
asyncpg
pgcli
passlib
python-jose[cryptography]
python-multipart
cachetools
python-dotenv
psycopg[binary,pool]
dependency-injector
